src/requirements.txt
selenium==3.141.0
pandas==1.0.1
requests==2.23.0
PyHamcrest==2.0.2